refactor(normalization): share error cause formatter
