+++ packages/normalization-core/src/error-coercion.ts
@@ -122,3 +122,21 @@ export function toErrorObject(value: unknown, fallbackMessage: string): Error {
   }
   return error;
 }
+
+/** Renders a non-Error cause as useful text without throwing. */
+export function stringifyNonErrorCause(value: unknown): string {
+  if (value === null) {
+    return "null";
+  }
+  if (typeof value === "string") {
+    return value;
+  }
+  if (typeof value === "number" || typeof value === "boolean" || typeof value === "bigint") {
+    return String(value);
+  }
+  try {
+    return JSON.stringify(value) ?? Object.prototype.toString.call(value);
+  } catch {
+    return Object.prototype.toString.call(value);
+  }
+}
